Delete HA network when last HA router is deleted
Currently when the last HA router of a tenant is deleted the HA network
belonging to this tenant is not removed. While running tempest aganist an
OpenStack setup where tenant VLANs (with small VLAN range) is used we hit
the limits are tempest tests start to fail as we cannot create new networks.
This patch addresses this issue by deleting the HA network when the last HA
router is deleted for the tenant.

Closes-Bug: #1367157

Co-Authored-By: Ann Kamyshnikova<akamyshnikova@mirantis.com>

Change-Id: I1d50b973aed4148857ac3d2bbee0d38e2e199783
(cherry picked from commit 853f7d7a74a2281b0cd34c337362e73635e501a2)